Replace Servant.Common.BaseUrl with Servant.Client.Core.Internal.BaseUrl
