+++ default.nix
@@ -1,7 +1,7 @@
-{ mkDerivation, aeson, base, bytestring, case-insensitive
-, containers, data-default, exceptions, ghcjs-dom, http-api-data
-, http-media, jsaddle, mtl, network-uri, reflex, reflex-dom
-, reflex-dom-core, safe, scientific, servant, servant-auth, stdenv
+{ mkDerivation, base, aeson, bytestring, case-insensitive, containers
+, data-default, exceptions, ghcjs-dom, http-api-data, http-media
+, jsaddle, mtl, network-uri, reflex, reflex-dom, reflex-dom-core, safe
+, scientific, servant, servant-auth, servant-client-core, stdenv
 , string-conversions, text, transformers
 }:
 mkDerivation {
@@ -10,7 +10,7 @@ mkDerivation {
   src = builtins.filterSource
         (path: type:
         baseNameOf path != "result"
-        && baseNameOf path != "nix" 
+        && baseNameOf path != "nix"
         ) ./.;
   configureFlags = [ "-fexample" ];
   isLibrary = true;
@@ -18,8 +18,8 @@ mkDerivation {
   libraryHaskellDepends = [
     base bytestring case-insensitive containers data-default exceptions
     ghcjs-dom http-api-data http-media jsaddle mtl network-uri reflex
-    reflex-dom-core safe servant servant-auth string-conversions text
-    transformers
+    reflex-dom-core safe servant servant-auth servant-client-core
+    string-conversions text transformers
   ];
   executableHaskellDepends = [
     aeson base reflex reflex-dom scientific servant text
