+++ servant-reflex.cabal
@@ -21,7 +21,6 @@ library
     Servant.Reflex.Multi
 
   other-modules:
-    Servant.Common.BaseUrl
     Servant.Common.Req
 
   hs-source-dirs: src
@@ -43,6 +42,7 @@ library
     safe                >= 0.3.9 && < 0.4,
     servant             >= 0.8  && < 0.15,
     servant-auth        >= 0.2.1 && < 0.4,
+    servant-client-core,
     string-conversions  >= 0.4  && < 0.5,
     text                >= 1.2  && < 1.3,
     transformers        >= 0.4  && < 0.6
